chore: remove comments about `Renderer2` in tests
This comment isn't really needed because if you try to move the provider,
the test will fail. Also, at least one of these comments is outdated, and
there is at least one place this comment is missing. I don't think it's
particularly valuable to try to keep these comments in sync with the code.

This is a backport of 2a4a1720cab5ecae1affe1abc364be198c9b7918.