chore: remove unused module-level `Renderer2` providers in tests
If needed, providers that depend on `Renderer2` should be provided at
the component injector to ensure a proper instance of `Renderer2` is
provided. This is especially important for services that register event
listeners and need to be destroyed properly.

This is backport of 34faa1a8d1fecd924b21708a0fd873e08fbf18d7.