feat(datagrid): add ctrl/command click row selection mode
This mode enables a row selection mode that works similar to the Kendo
datagrid and Windows File Explorer. If you click a row, all other rows
are deselected unless you press ctrl or command when clicking the row.

This feature is implemented a separate directive because it's easier to
do it this way than modifying the datagrid code directly. The row
selection is toggled via a property setter which means the `MouseEvent`
is not available to check if a modifier key was pressed.

Co-Authored-By: Steve Haar <info@stevehaar.com>