chore: auto accept merged visual changes
When a PR is rebased and merged, the SHA of the commit that lands in
the base branch is different than the SHA of the commit in the PR. This
makes Chromatic report the same visual change in subsequent PRs because
it uses the incorrect baseline build.

The Chromatic GitHub app is supposed to fix this, but it doesn't.

This commit reverts b73a27a6561ddb0cc6d0635215db6a304279bdc4 and
reapplies f4fb2ad0cdd865bf6531f0cf53b3707c1396cb0b.