chore: auto accept merged visual changes
When a PR is rebased and merged, the SHA of the commit that lands in
the base branch is different than the SHA of the commit in the PR. This
makes Chromatic report the same visual change in subsequent PRs because
it uses the incorrect baseline build.

I think we can remove this once we get the Chromatic GitHub app approved
since Chromatic should be able to track baselines properly. But this
workaround should work until then.