fix: ensure validation icons are the correct color
The icon `status` attribute sets the icon color and fill which could
result in slightly different colors being used for the validation icon
and text. I think it's best to keep the icon `status` attributes for
semantics/intention simply set the color and fill on the validation
icons to the correct form color.

closes #305

This is a backport of 28bbc624a27d7e58896d2b672bf324a175d87843.