fix(tooltip): keep tooltip content visible as user hovers
Previously, the tooltip content was hidden as soon as the user moused
out of the trigger.

Now, the tooltip content will stay visible if the user mouses from the
trigger into the content. To support this, the tooltip content has an
expanded hover target provided by the `::after` pseudo element.

VPAT-617