fix(datagrid): show full border when detail pane is open
The `display: block` was interfering with the flex styles that make
the placeholder element below the rows take up the remaining vertical
space. This cased the border between the rows and the detail pane not
to extend past the rows.

This fixes a regression that was introduced in 086f45b2e87bfd32b621b3b51a9642e59f6ac1cf.