feat(wizard): wip: redesign wizard for accessibility
VAT-701 and VPAT-776

Note: The `clrModalInitiallyFocusTitle` input was needed so that the
wizard for initially-focus only part of the modal title. This input is
not intended for external use. We should refactor the modal to extract
out a bare bones `clr-internal-modal` that has no prescribed content
structure. This would make it so we don't have to add things to `clr-modal`
just to support the wizard.

BREAKING CHANGE: This is visually-breaking change for the wizard. The
wizard modal now as a full-width title and the step nav was redesigned.
The DOM structure was significantly altered.