feat(wizard): rename "sidenav" css custom properties to "stepnav"
BREAKING CHANGE: There were css custom properties for the wizard that
used "sidenav" in the name. This is inconsistent with the rest of the
code which uses "stepnav" for the wizard nav element, so `sidenav`
was replaced with `stepnav` in all such variables.

The same renames were made for the scss variables (prefixed with `$`).