fix(forms): update validation with onpush strategy components
closes #476

The "touched" control check was nice to have but not required for the proper validation operation. Also, the event order is not stable in different Angular versions and orders of module imports. The setTimeout() was trying to ensure the correct order. But it leads to regressions with OnPush components.
This fix removes the setTimeout at the expense of sacrificing the "touched" check.

Signed-off-by: Ivan Donchev <idonchev@vmware.com>