fix(datagrid): add workaround to enable `clrDgSelectable` with `*ngFor`
It is impossible (or very difficult) to grab the correct `trackBy`
function via the row iterator structural directive. The `ngForTrackBy`
directive would grab the `trackBy` function from any `*ngFor` in the
datagrid, not just the row iterator.

Further, needing the item index to track selection introduces a race
condition when `*ngFor` is used. We don't have the collection of items
when the rows inputs are set, so `clrDgSelectable` doesn't work.

The fix is to pass a "track by" function directly to the datagrid. This
change introduces a `clrDgItemsTrackBy` input that will used if provided.
Otherwise, the datagrid will fall back to the previous behavior of
trying to grab the `trackBy` function from the row iterator structural
directive. The fallback behavior is deprecated and scheduled for removal
in v17.

closes #157