fix(datagrid): prevent "manage columns" from making rows selectable
This is a workaround to prevent the items `trackBy` function from
being replaced when the "manage columns" button is clicked. This is
not a complete solution. If there is another `ngForTrackBy` function
within the datagrid in application code, it could sill replace the
items `trackBy` function whether it is the row iterator or not.

closes #475
closes #503