fix(datagrid): prevent "manage columns" from making rows selectable
This is a workaround to prevent the items `trackBy` function from
being replaced when the "manage columns" button is clicked. This is
not a complete solution. If there is another `ngForTrackBy` function
within the datagrid in application code, it could sill replace the
items `trackBy` function whether it is the row iterator or not.

This is a backport of 32710a35787a7c2f24269452c81fd4511d7521d7 to 13.x.

closes #475
closes #503