feat(datagrid): use only `clrDgItemsTrackBy` for selection tracking
I removed two tests that appeared to test an optimization to skip work
if there are no items. These optimizations are invalid because it may
appear that the datagrid has no items when it just doesn't have items yet.

BREAKING CHANGE: The row iterator `trackBy` function will no longer be
used for selection tracking. Update your code to pass `clrDgItemsTrackBy`
to the `clr-datagrid` component if you pass `trackBy` to the row iterator.