fix(forms): fix template-driven forms validation regression (#548)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [x] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)
- [ ] If applicable, have a visual design approval

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Documentation content changes
- [ ] Other... Please describe:

## What is the current behavior?

Forms validation regressed multiple times, after certain Angular update,
after which the order of imports of Angular's FormsModule and
ClarityModule started to matter for the order of receiving `touched`
event and angular validation status event.
Previously, we were trying to augment our internal validity state, based
on whether `touched` had arrived or not. With the race condition
described above, this was causing different side effects, depending on
the fix attempted, including:
- form elements only got validated on the second `blur`
- form elements got validated immediately, when page was opened
Additional problem was that `touched` changes do not have an event
handler in the Angular Control object, which makes manual
synchronization in the above pattern nearly impossible.

Issue Number: #511 

## What is the new behavior?

Validity is now updated only after the first `blur`  on the element.
We are no longer augmenting the internal validity state, which was
making it different from the Angular validity state.
Now both Clarity and Angular validity states are in sync.
The decision. whether to display the element as valid, or not yet
validated, based on the `touched` state is not handled in the control
state service, but in the Form component itself, which allows us to use
the standard Angular data bindings to actively monitor both validation
state and `touched` state. With this approach, the elements validity
classes get updated regardless of the order in which validity and
`touched` state changes arrive.

## Does this PR introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

## Other information

Due to the complex nature of the setup in which these issues are
observed, we are aiming for them to be covered by future visual
regression testing.
For this purpose, two storybook pages are defined - for reactive and for
template-driven setup.
So far, I am unable to create a story with OnPush change detection in
storybook, so I have added two more pages to the Clarity demo app, for
the OnPush scenarios.

I recommend a manual testing procedure to be run on the following pages
in the Dev App:
- Template Driven
- Template Driven (OnPush)
- Reactive
- Reactive (OnPush)

The procedure would require testing that:
- No field on the pages is pre-validated. 
- The field marked as `required` gets validated on the first `blur`
event on the element.
- The field marked as `required` does not need a second `blur` event to
get validated.

This will need to get back ported.