feat(checkbox,radio,tree): refactor checbox/radio label styles to class (#568)
Please check if your PR fulfills the following requirements:

- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)
- [ ] If applicable, have a visual design approval

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

- [ ] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [x] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Documentation content changes
- [ ] Other... Please describe:

The checkbox/radio label styles are designed to only work with the
semantic label element.

Issue Number: N/A

This change refactors the checkbox/radio label styles to use a class, as
a valid use case for the selectable rows feature in data grid causing a
label to be nested within a label. In the future, we may want to expand
this to be an option on the checkbox/label component, since another
valid use case for a checkbox is to have it be a wrapping element around
a checkbox.

- [ ] Yes
- [x] No

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->