chore: enable Angular's strict checks in library and demo app (backport to 13.x) (#597)
The `FocusTrap` is used as a base class. It is not injectable.

This change required minor changes to templates in the demo app.

The `strictInputAccessModifiers` option is implied by `strictTemplates`,
so I removed that option.

This is a backport of 2019766b9980b30735a8667e89922e99fc1ae41d to 13.x.