fix(combobox): prevent "expression changed" errors after modal close
When focus is returned to a combobox after a modal is closed, the
combobox would throw a "expression has changed after it was checked"
error for the `[class.clr-focus]` binding.

I am not sure why the error is thrown because the expression's value is
changed from `false` to `true` on focus, but running change detection
for the combobox immediately after it receives focus fixes the issue.

closes #663