fix(datagrid): re-map selection if row item references are changed (#672)
Previous behavior:

If the row item references are changed but the row component instances
are not changed, the selection was not re-mapped to the new item
instances.

This was because the selection was only re-mapped if the rows
`QueryList` emitted a change. If `[ngForTrackBy]` is used and the items
collection is updated to new instances of the same items, the selection
was not mapped to the new item instances. This caused the datagrid to
display an inconsistent/incorrect view (e.g. a non-zero selection count
but no rows actually checked).

New behavior:

If the row item references are changed, the selection is re-mapped to
the new object references even if the rows `QueryList` does not emit a
change.

closes #671