fix(datagrid): re-map selection if row item references are changed (#682)
Previous behavior:

If the row item references are changed but the row component instances
are not changed, the selection was not re-mapped to the new item
instances.

This was because the selection was only re-mapped if the rows
`QueryList` emitted a change. If `[ngForTrackBy]` is used and the items
collection is updated to new instances of the same items, the selection
was not mapped to the new item instances. This caused the datagrid to
display an inconsistent/incorrect view (e.g. a non-zero selection count
but no rows actually checked).

New behavior:

If the row item references are changed, the selection is re-mapped to
the new object references even if the rows `QueryList` does not emit a
change.

closes #671

This is a backport of 701bc436c48a9c0c3f8305076a339642be91f42f to 13.x.