feat(popover): use scroll value in alignment and visible event added (#775)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)
- [ ] If applicable, have a visual design approval

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

- [ ] Bugfix
- [X] Feature
- [ ] Code style update (formatting, local variables)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Documentation content changes
- [ ] Other... Please describe:

## What is the current behavior?

Currently if you open a popover it's being rendered at the top left
corner with opacity 0 which just makes it invisible but it's available
for interactions. So when you click the action button in datagrid and
you have available scroll the page gets scrolled to the top due to
focusing of the first item in that popover which is positioned in the
top corner.

Issue Number: CDE-150

## What is the new behavior?

We are using the scrolled value in the calculations because while the
content was scrolled to the top the wrong positioned popover was not
noticed. Using an event which emits once the popover is visible fixes a
lot of possible and current issues because of most components are
awaiting popover open event which is not when the popover is really
ready for interactions.
Initially I fixed the issue with using 'visibility: hidden' instead of
opacity to hide the popover but that creates a lot of other issues both
in tests and other places which in perfect scenario should be fixed but
in the current moment the task gets out of scope a lot.

## Does this PR introduce a breaking change?

- [ ] Yes
- [X] No

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

## Other information