fix(tooltip) allow user to interact with tooltip content using mouse
The previous change to introduce a fix for VPAT-617 meant the user was
unable to copy text out of a tooltip because the ::after content
overlays the tooltip content.

This change adds a -1 z-index to the ::after so that it is underneath
which exposes the tooltip content div to allow text to be selected and
copied.