fix(datagrid): change pagination property types to `number`
These properties always return a `number` value and ignore non-`number`
values in the setters. The `string | number` type causes consumers to
have to unnecessarily assert that the value is a number when using these
properties.

If this change results in a build error, it would be due code that was
previously incorrect and non-functioning since `string` values were
ignored by these setters. Such values are still ignored, but the compiler
may complain now.