fix(stack-view): ignore input events in stack block (#841)
The stack block handles the space and enter keys to toggle expansion. We
don't want the stack block to expand or collapse on click, space, or
enter inside of an input.

The stack block prevents default on `keydown.space` which prevents the
space from being "typed" into the input. This change fixes that as well.

closes #710