fix(stack-view): ignore input events in stack block
This is a backport of b3fc2c6cb82ed7bdb53f474cbcc8553c0894358b (#841) to 13.x.

The stack block handles the space and enter keys to toggle expansion. We
don't want the stack block to expand or collapse on click, space, or
enter inside of an input.

The stack block prevents default on `keydown.space` which prevents the
space from being "typed" into the input. This change fixes that as well.

closes #710