chore(vrt): change filter to exclude default stories (#936)
Updates the VRT screenshot tests from only taking screenshots of stories
ending with `--variants` to taking screenshots of every story except
those ending in `--default`. As stories are converted to use the
standard CSFv3 syntax, we'll end up with more individual stories but for
those that are converted, we won't have the --variants version of the
page.