fix(vertical-nav): fix inconsistent padding and font-weight of nav item groups (#1018)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)
- [x] If applicable, have a visual design approval

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Documentation content changes
- [ ] Other... Please describe:

## What is the current behavior?

In the default v16 theme, a top-level anchor tag category item will
render indented and not visually aligned with the other top-level list
items. The font-weight of the top-level anchor tag items is also
incorrect, as all top-level groups should use the 600 font-weight.

Issue Number: CDE-1349

## What is the new behavior?

The top-level category items are rendered in the appropriate weight
(both buttons and links). An additional ticket has been created to
investigate if the vertical nav should support both buttons and links
(angular router works with both, and links have problematic a11y issues
associated with them in this context). This change also fixes the
misalignment between links and buttons as top-level category items.

For the "old" theme, this change changes the top-level nav group link
items to use the same heavier font-weight as buttons in the same
context. Since this change updates the anchor tag top-level nav group
items to match the buttons used in the same context, I consider it to be
an acceptable compromise for the old theme since that theme will be
removed in the upcoming v17.

## Does this PR introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

## Other information

---------

Co-authored-by: GitHub <noreply@github.com>