feat(wizard): add back inline wizards in 16.x
This is a port of c5d3f09d50af56c27f50476c88c0a0816b96d1e6 (#1067) to 16.x.

This reverts commit 7470769480b843dc332adbd25f67624ce4979e56 (#419),
removes the deprecated comments, and adds a comment explaining when
inline wizards are used.

Inline wizards are used by vSphere plugins to display a wizard inside an iframe
that is inside a modal.

CDE-154