chore: add retries to playwright to prevent flaky tests (#1089)
Update the playwright config to allow two retries for failed tests. When
a test fails, that individual test will be retried twice, and if it
fails every time, the last screenshot will be retained for different
comparisons. This should help overcome the flaky button issues seen on
this PR and #1088 .

---------

Co-authored-by: GitHub <noreply@github.com>