fix(select): add select option background-color (port of #1088)
Please check if your PR fulfills the following requirements:

- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)
- [ ] If applicable, have a visual design approval

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Documentation content changes
- [ ] Other... Please describe:

The background of the select options dropdown is white when in dark mode
(on Windows), making the text difficult to read.

Issue Number: CDE-1492

This fix adds the background-color of the options to match the textarea
in the clr theme, and uses the cds-alias-object-interaction-background
token in the cds theme.

- [ ] Yes
- [x] No

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

![image
(2)](https://github.com/vmware-clarity/ng-clarity/assets/34519388/25dc547e-5d03-4981-b49b-219c8d26c727)