chore: add retries to playwright to prevent flaky tests (port of #1089) (#1096)
Update the playwright config to allow two retries for failed tests. When
a test fails, that individual test will be retried twice, and if it
fails every time, the last screenshot will be retained for different
comparisons. This should help overcome the flaky button issues seen on
this PR and #1088 .

Port of #1089

Co-authored-by: GitHub <noreply@github.com>