chore: delete visual snapshot update branch immediately
There was code to cleanup visual snapshot update branches if no changes,
but that job only runs after after failed build.

This resulted in visual snapshot update branches persisting after
successful builds.

We can immediately delete the visual snapshot branches because we just
need the commit in the repo. The branch is not needed after that.