fix(button-group): handle keyboard focus for async menu buttons
This is a port of fa1da8a748f5a63c2acabadb66e6b0d8ed5511b0 (#1142) to 17.x.

With both synchronous and asynchronous population, focus is moved to the
first menu button when the button group menu is opened, keyboard focus
works properly, and the menu is closed when one of the menu buttons is
activated.

CDE-1571
closes #952