fix(card): resolve clickable card hover shifting text (#1166)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)
- [ ] If applicable, have a visual design approval

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

- [X] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Documentation content changes
- [ ] Other... Please describe:

## What is the current behavior?

When hovering over clickable card the border width animates from 1px to
2px, causing the text content contained within the card to noticeably
shift. This can be seen when hovering the clickable card on our
[website](https://clarity.design/documentation/card/code#clickable-cards)
and our
[storybook](https://storybook.clarity.design/?path=/story/card-card--default).

Issue Number: CDE-917

## What is the new behavior?

This changes the clickable border-width token to 1px (which is the
correct value for non-hovered clickable cards), and then uses the
box-shadow to set a 1px shadow around the element. Since both the
box-shadow and the border-color are the same color, and they're both 1px
by default, the result is a non-shifting 2px stroke on hover.

## Does this PR introduce a breaking change?

- [ ] Yes
- [X] No

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

## Other information