fix(accordion): header spacing and disabled panel content color (#1168)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)
- [ ] If applicable, have a visual design approval

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Documentation content changes
- [ ] Other... Please describe:

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

Issue Number: CDE-1603

## What is the new behavior?

## Does this PR introduce a breaking change?

- [x] Yes
- [ ] No

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

## Other information

This PR introduces changes in the API (removed obsolete custom
properties) of the accordion / stepper and the appearance of the
stepper.

There are spacing differences (hence it was marked as breaking change)
and color differences. Previous version was not following figma specs or
was not well defined, so this is marked as a "fix".

Removed obsolete custom property is:
--clr-accordion-header-button-padding

Spacing changes are:

-we used to have stepper header panel padding set to 18px on top and
bottom, and the overall height of the header panel was calculated to
56.4px; overall height is calculated to exactly 56px;
-when a step is completed header panel used to have bigger height (60px)
because of the size of the complete icon, now it stays to 56px on
complete/error etc;
-gap between header panel container children was changed to 24px to
reflect figma spec;
-there was smaller padding-left of the header panel container and some
margin-left to the first child; changed to bigger padding-left and no
margin for the child -> this helps with vertical alignments when there
are two rows in smaller screens;
-top and bottom padding for header container was reduced from 18 to 8px
for "breathing" space when there is a very long text or a small screen;


Preview link look in normal screen (padding right is 24 and padding left
is 32 because of the box-shadow blue indicator which is 8px):


![image](https://github.com/vmware-clarity/ng-clarity/assets/132376042/c7b9aa8b-7571-4acb-a0fc-278bfe02efcf)

Smaller screen (big enough to have only one row though):


![image](https://github.com/vmware-clarity/ng-clarity/assets/132376042/7c8a3ed2-1ac6-41a8-8ff5-c1af339fe19a)

When we have two rows (they are now vertically aligned, the gap is 24px
by design and that's why we have so big space between the rows):


![image](https://github.com/vmware-clarity/ng-clarity/assets/132376042/ee5002a3-2639-45ac-996a-6fed29b0b8cf)

When texts are long (there is this "breathing" space and texts are not
fixed to the header container border:


![image](https://github.com/vmware-clarity/ng-clarity/assets/132376042/10615561-d5ab-4e37-9b8b-2a1fce78d50c)




BREAKING CHANGE: Changed spacing in stepper header

---------

Co-authored-by: GitHub <noreply@github.com>