fix(modal): prevent error when modal is escaped with a focused input
If you have an input inside a modal (specifically an input with `ngModel`)
and escape the modal with the input focused, an "expression change after
it was checked" error might be thrown.

I think this is due to Angular updating the input's touched status after
the element has been destroyed due to modal being closed.

Blurring (unfocusing) the active element before closing the modal solves
this issue. I think this is safe fix because the modal's trigger element
is focused immediately after the modal is closed so focus is not lost.

CDE-1662