fix(datagrid): fix pressed action overflow icon color (#1276)
Please check if your PR fulfills the following requirements:

- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)
- [ ] If applicable, have a visual design approval

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Documentation content changes
- [ ] Other... Please describe:

When the action overflow button is in a pressed state, the icon color
shifts to a lighter tone background color and the icon appears to
disappear due to a lack of color contrast with the hover background
color. this is a result of the icon color being mapped to a background
color in the system, when it should be mapped to color-active.

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

Issue Number: CDE-1689

The icon does not appear to disappear when the user is pressing the
button.

- [ ] Yes
- [x] No

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->